Check answers to HW 8
Question 1 (b) v. The question says "if you answer 'yes' to (iii)", not
to iv, so I should respond with the reason why it is not conflict
serializable, not why it is not view serializable.

Question 3
(c) Although my answer is accepted, I'd rather adopt the other answer in
the solution, which correctly handles the case when a page is inserted
at the end of a page.
(d) A shared lock is required to read the start_city attribute of each
row! The shared lock also makes page-level locks obsolete.
(e) I thought we shouldn't lock the entire database, because
what if you want to create tables in the database at the same time? This
 is wrong, because "Delete ALL the records from ALL tables" means that
 any newly created tables must be also updated in the query. We have to
lock the entire database.